Finished GPS Chase Car code.
Picked up the chase car gps code again; this is a complete redraft, and
it would appear to work! By work, I mean, it parses gps strings
correctly, the restarting and reconfiguring (signals and pthread_cond)
work perfectly, however I haven't tested it with a real gps since I
don't have one wired up at the moment. Some other things to test include
plugging in a dodgy gps or opening a totally responseless port to check
that it will still EINTR-out and reconfigure to a new port when the user
wants.